V1 completed
Some new bells and whistles.

Use the keyboard shortcut 'p' to set the first "scale point".  Then navigate so that the lower right of the image is visible and hit 'p' again.  This will determine the resize ratio to shrink the image so that the entire region that you just panned through is visible within a single PIXEL_WIDTH x PIXEL_HEIGHT area.